move buttons stuff from marcRecordUI.js to editorButtons.js (new file)
